+++ src/structlog/stdlib.py
@@ -263,7 +263,7 @@ class PositionalArgumentsFormatter(object):
         self.remove_positional_args = remove_positional_args
 
     def __call__(self, _, __, event_dict):
-        args = event_dict.get('positional_args')
+        args = event_dict.get("positional_args")
 
         # Mimick the formatting behaviour of the stdlib's logging
         # module, which accepts both positional arguments and a single
@@ -272,9 +272,9 @@ class PositionalArgumentsFormatter(object):
         if args:
             if len(args) == 1 and isinstance(args[0], dict) and args[0]:
                 args = args[0]
-            event_dict['event'] = event_dict['event'] % args
-            if self.remove_positional_args:
-                del event_dict['positional_args']
+            event_dict["event"] = event_dict["event"] % args
+        if self.remove_positional_args and args is not None:
+            del event_dict["positional_args"]
         return event_dict